Fixed the tests and deprecation warnings.

+++ db/schema.rb
@@ -90,8 +90,8 @@ ActiveRecord::Schema[7.1].define(version: 2025_11_30_015033) do
   end
 
   create_table "order_products", force: :cascade do |t|
-    t.bigint "product_id", null: false
-    t.bigint "order_id", null: false
+    t.integer "product_id", null: false
+    t.integer "order_id", null: false
     t.string "size"
     t.integer "quantity"
     t.datetime "created_at", null: false
@@ -126,7 +126,7 @@ ActiveRecord::Schema[7.1].define(version: 2025_11_30_015033) do
     t.string "name"
     t.text "description"
     t.integer "price"
-    t.bigint "category_id", null: false
+    t.integer "category_id", null: false
     t.boolean "active"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
@@ -149,7 +149,7 @@ ActiveRecord::Schema[7.1].define(version: 2025_11_30_015033) do
   create_table "stocks", force: :cascade do |t|
     t.string "size"
     t.integer "stock_level"
-    t.bigint "product_id", null: false
+    t.integer "product_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.integer "price"
